Fix layering of 3d structure viewer and top bar
The ProtVista protein structure viewer on the target profile page
rendered over the logo bar when scrolling past it, seemingly because the
logo bar existed in the DOM inside the local stacking context of another
app bar with a z-index value lower than 40001.

This still leaves the issue that the bar also overlays the 3d structure
viewer when it's expanded to fill the viewport, blocking some of the
buttons of the viewer.